Rename index.js to index.html
The index.html file located in the public folder of a newly created React app serves as the entry point for your web application. By default, it includes:

A <div> element with the ID root, which is where your React components will be rendered.
Basic HTML structure including <head> and <body> tags.
Links to your CSS files and meta information.
This file is crucial for loading your React application and providing a base HTML structure that React builds upon.